Refactor query parameter button helpers to reduce duplication
Extract common query parameter preservation logic from AttributeTypeAdminButtonHelper,
CategoryAdminButtonHelper, CommonCategoryAdminButtonHelper, and ReportAdminButtonHelper
into a new QueryParameterButtonHelper base class in admin_site/wagtail.py. This eliminates
duplicate code across multiple admin modules while preserving the same functionality.
